+++ app/soapbox/jest/test-helpers.tsx
@@ -4,6 +4,7 @@ import { render, RenderOptions } from '@testing-library/react';
 import { renderHook, RenderHookOptions } from '@testing-library/react-hooks';
 import { merge } from 'immutable';
 import React, { FC, ReactElement } from 'react';
+import { Toaster } from 'react-hot-toast';
 import { IntlProvider } from 'react-intl';
 import { Provider } from 'react-redux';
 import { MemoryRouter } from 'react-router-dom';
@@ -15,7 +16,6 @@ import { ChatProvider } from 'soapbox/contexts/chat-context';
 import { StatProvider } from 'soapbox/contexts/stat-context';
 import { queryClient } from 'soapbox/queries/client';
 
-import NotificationsContainer from '../features/ui/containers/notifications-container';
 import { default as rootReducer } from '../reducers';
 
 import type { AnyAction } from 'redux';
@@ -59,7 +59,7 @@ const TestApp: FC<any> = ({ children, storeProps, routerProps = {} }) => {
               <IntlProvider locale={props.locale}>
                 {children}
 
-                <NotificationsContainer />
+                <Toaster />
               </IntlProvider>
             </ChatProvider>
           </QueryClientProvider>