+++ app/soapbox/jest/test-setup.ts
@@ -1,5 +1,7 @@
 'use strict';
 
+import { toast } from 'react-hot-toast';
+
 import { __clear as clearApiMocks } from '../api/__mocks__';
 
 // API mocking
@@ -15,6 +17,11 @@ require('fake-indexeddb/auto');
 
 // Mock external dependencies
 jest.mock('uuid', () => ({ v4: jest.fn(() => '1') }));
+
+// Clear toasts after each test.
+afterEach(() => {
+  toast.remove();
+});
 
 const intersectionObserverMock = () => ({ observe: () => null, disconnect: () => null });
 window.IntersectionObserver = jest.fn().mockImplementation(intersectionObserverMock);